Conflicts:
	Implementation/MISServices/CompanyService.cs

+++ MPC.Web/Areas/Api/Models/PaymentGateway.cs
@@ -6,7 +6,7 @@
         public string BusinessEmail { get; set; }
         public string IdentityToken { get; set; }
         public bool IsActive { get; set; }
-        public int? CompanyId { get; set; }
+        public long? CompanyId { get; set; }
         public int? PaymentMethodId { get; set; }
         public string SecureHash { get; set; }
     }
